move models folder to shared
