+++ Src/Campus.Master.API/UI/src/app/data/dao/impl/CategoryDAOArray.ts
@@ -1,5 +1,5 @@
 import { CategoryDAO } from '../interface/CategoryDAO';
-import { Project } from '../../../model/Project';
+import { Project } from '../../../shared/models/project';
 import { Observable, of } from 'rxjs';
 import { TestData } from '../../testData';
