+++ Src/Campus.Master.API/UI/src/app/data/dao/impl/StatusDAOArray.ts
@@ -1,7 +1,7 @@
 import { Observable, of } from 'rxjs';
 import { TestData } from '../../testData';
 import { StatusDAO } from '../interface/StatusDAO';
-import { Status } from '../../../model/status';
+import { Status } from '../../../shared/models/status';
 
 export class StatusDAOArray implements StatusDAO {
