+++ Src/Campus.Master.API/UI/src/app/data/dao/interface/PriorityDAO.ts
@@ -1,5 +1,5 @@
 import { CommonDAO } from './CommonDAO';
-import { Priority } from '../../../model/priority';
+import { Priority } from '../../../shared/models/priority';
 
 export interface PriorityDAO extends CommonDAO<Priority> {
 }
