+++ Src/Campus.Master.API/UI/src/app/data/dao/interface/StatusDAO.ts
@@ -1,5 +1,5 @@
 import { CommonDAO } from './CommonDAO';
-import { Status } from '../../../model/status';
+import { Status } from '../../../shared/models/status';
 
 export interface StatusDAO extends CommonDAO<Status> {
