+++ Src/Campus.Master.API/UI/src/app/shared/models/status.ts
