+++ Src/Campus.Master.API/UI/src/app/views/tasks_board/task_board_left/project-tasks.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { Task } from '../../../model/task';
+import { Task } from '../../../shared/models/task';
 
 @Component({
   selector: 'app-project-tasks',
